after power measurements

+++ VivadoHLS/source/test_AES_full.cpp
@@ -1,14 +1,11 @@
 #include "AESfunctions.h"
 
-void AES_Full(bool cipher_or_i_cipher, unsigned char Nr,
-		unsigned char data_in[stt_lng], unsigned char data_out[stt_lng]);
+void AES_Full(bool cipher_or_i_cipher, unsigned char Nr, unsigned char data_in[stt_lng], unsigned char data_out[stt_lng]);
 
 void KeyExpansion(unsigned char* inputKey, unsigned short Nk, unsigned char* expandedKey);
 
-
-int main()
-{
-	unsigned short NkVals[3] = {4, 6, 8};
+int main() {
+	unsigned short NkVals[3] = { 4, 6, 8 };
 	unsigned short Nk, CipherKeyLenghth, Nr, ExtdCipherKeyLenghth;
 
 	// create a dummy test cipher key
@@ -34,16 +31,13 @@ int main()
 		Nr = (Nk > Nb) ? Nk + 6 : Nb + 6; // = 10, 12 or 14 rounds
 		ExtdCipherKeyLenghth = (Nr + 1) * stt_lng; // bytes in the expanded cipher key
 
-		cout << "AES with Nb = " << Nb << " columns, Nk = " << Nk << " (32-bit) words i.e. CipherKeyLenghth = "
-			<< CipherKeyLenghth << " bytes (or " << CipherKeyLenghth * 8 << " bits), Nr = " << Nr << " rounds" << endl << endl;
-
+		cout << "AES with Nb = " << (unsigned short) Nb << " columns, Nk = " << Nk << " (32-bit) words i.e. CipherKeyLenghth = " << CipherKeyLenghth << " bytes (or " << CipherKeyLenghth * 8 << " bits), Nr = " << Nr << " rounds" << endl << endl;
 
 //		cout << "key = ";
 //		for (unsigned short i = 0; i < CipherKeyLenghth; i++) {
 //			key[i] = (unsigned char)i;
 //			cout << "0x" << hex << (unsigned short)key[i] << " ";
 //		} cout << endl << endl;
-
 
 //		KeyExpansion(key, Nk, expandedKey);
 //		cout << "expandedKey = ";
@@ -52,29 +46,34 @@ int main()
 //		for (unsigned short i = 0; i < ExtdCipherKeyLenghth; i++) { cout << "0x" << hex << (unsigned short)expandedKey[i] << " "; }
 //		cout << endl << endl;
 
-
 //		cout << "plaintext = ";
 //		for (unsigned short i = 0; i < stt_lng; i++) { cout << plaintext[i] << " "; }
 //		cout << "<=> ";
 //		for (unsigned short i = 0; i < stt_lng; i++) { cout << "0x" << hex << (unsigned short)plaintext[i] << " "; }
 //		cout << endl << endl;
 
-
 		// encrypt
 		AES_Full(true, (unsigned char) Nr, plaintext, ciphertext);
 		cout << "ciphertext = ";
-		for (unsigned short i = 0; i < stt_lng; i++) { cout << dec << (unsigned short)ciphertext[i] << " "; }
+		for (unsigned short i = 0; i < stt_lng; i++) {
+			cout << dec << (unsigned short) ciphertext[i] << " ";
+		}
 		cout << "<=> ";
-		for (unsigned short i = 0; i < stt_lng; i++) { cout << "0x" << hex << (unsigned short)ciphertext[i] << " "; }
+		for (unsigned short i = 0; i < stt_lng; i++) {
+			cout << "0x" << hex << (unsigned short) ciphertext[i] << " ";
+		}
 		cout << endl << endl;
-
 
 		// decrypt
 		AES_Full(false, (unsigned char) Nr, ciphertext, decrypted_plaintext);
 		cout << "decrypted_plaintext = ";
-		for (unsigned short i = 0; i < stt_lng; i++) { cout << decrypted_plaintext[i] << " "; }
+		for (unsigned short i = 0; i < stt_lng; i++) {
+			cout << decrypted_plaintext[i] << " ";
+		}
 		cout << "<=> ";
-		for (unsigned short i = 0; i < stt_lng; i++) { cout << "0x" << hex << (unsigned short)decrypted_plaintext[i] << " "; }
+		for (unsigned short i = 0; i < stt_lng; i++) {
+			cout << "0x" << hex << (unsigned short) decrypted_plaintext[i] << " ";
+		}
 		cout << endl << endl;
 
 		if (plaintext[0] == decrypted_plaintext[0])
